added update mechanism, DO NOT MERGE b4 seeing commit notes!
this "update mechanism" only deletes old caches. old means caches that
are caches with names other than the name in cache_name.txt. It is
important not to cache cache_name.txt and always serve it with HTTP
headers that allow no HTTP level caching. Update checking occurs every
300 seconds and if a new cache is released and sw.js is not reinstalled,
this timer will clear the cache.

the old update method did not work since it relied on sw being reloaded
often, which is not the case on iPhone devices, where Safari runs almost
the whole uptime of the phone and it's not common for people to
shut down their precious little iPhones.

it worked on android as android is known for killing apps and chrome is
no exception.